move groups resources to their own file
